requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
crewai==0.28.8
crewai-tools==0.4.26
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
ollama==0.1.7
python-dotenv==1.0.0
langchain-community>=0.0.13
langchain>=0.1.10

# MongoDB integration
pymongo==4.6.0
aiofiles==23.2.1

# Azure integration
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
azure-storage-blob==12.19.0

# Monitoring and logging
opencensus-ext-azure==1.1.13
opencensus-ext-flask==0.8.0
opencensus-ext-logging==0.1.1

# Security and networking tools
requests==2.31.0
dnspython==2.4.2
paramiko==3.3.1
scapy==2.5.0
impacket==0.11.0
python-nmap==0.7.1

# Data processing
pandas==2.1.4
numpy==1.25.2
beautifulsoup4>=4.12.3
lxml==4.9.3

# Async support
asyncio==3.4.3
aiohttp==3.9.1

# Development and testing
pytest>=8.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# UI and templating
click>=8.1.0
rich>=13.0.0
typer>=0.9.0
jinja2>=3.1.0
